# src/main/resources/static/swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Lunit service APIs",
        "description": "Test project for Lunit Assignment",
        "contact": {
            "name": "Bulgan Galbadrakh",
            "email": "bulgan.galbadrakh@gmail.com"
        }
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Local ENV"
        }
    ],
    "paths": {
...
        "/users/{userId}/grids": {
            "get": {
                "tags": [
                    "grid-controller"
                ],
                "summary": "Get all existing grids of one user",
                "operationId": "findAllGrids",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "slideId",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 3
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users/{userId}/grids/{gridId}": {
            "get": {
                "tags": [
                    "grid-controller"
                ],
                "summary": "Get an one specific grid of one user",
                "operationId": "getOneGridOfUser",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "gridId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Grid"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Grid": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "intratumoralMin": {
                        "type": "number",
                        "format": "double"
                    },
                    "intratumoralAvg": {
                        "type": "number",
                        "format": "double"
                    },
                    "intratumoralMax": {
                        "type": "number",
                        "format": "double"
                    },
                    "stromalMin": {
                        "type": "number",
                        "format": "double"
                    },
                    "stromalAvg": {
                        "type": "number",
                        "format": "double"
                    },
                    "stromalMax": {
                        "type": "number",
                        "format": "double"
                    },
                    "userId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "slideId": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            }
        }
    }
}